+++ inf-terraform-azure/files/inputs2outputs.tf
@@ -6,6 +6,7 @@ output "inputs2outputs" {
   description = "all inputs passed to outputs"
   value = [{
     is_test          = var.is_test
+    location         = var.location
     meta_environment = var.meta_environment
     name             = var.name
   }]